updated error handling user controller

+++ src/api/controllers/user.controller.ts
@@ -7,8 +7,13 @@ export class UserController {
     public static async getUsers(req: Request, res: Response) {
         try {
             const response = await UserService.getUsers();
-            logger.log('info', 'api-UserController-getUsers() | SUCCESS')
-            res.status(200).send(response);
+            if (response) {
+                logger.log('info', 'api-UserController-getUsers() | SUCCESS')
+                res.status(200).send(response);
+            } else {
+                logger.log('info', 'api-UserController-getUsers() | Error | No users found')
+                res.status(404).send({error: 'No users found'});
+            }
         } catch (e: unknown) {
             const typedE = e as Error
             logger.log('error', 'api-UserController-getUsers() | Error | ' + String(typedE.message))
@@ -19,8 +24,13 @@ export class UserController {
         try {
             const { id } = req.params;
             const response = await UserService.getUserById(id);
-            logger.log('info', 'api-UserController-getUserById() | SUCCESS')
-            res.status(200).send(response);
+            if (response) {
+                logger.log('info', 'api-UserController-getUserById() | SUCCESS')
+                res.status(200).send(response);
+            } else {
+                logger.log('info', 'api-UserController-getUserById() | Error | User not found')
+                res.status(404).send({error: 'User not found'});
+            }
         } catch (e: unknown) {
             const typedE = e as Error
             logger.log('error', 'api-UserController-getUserById() | Error | ' + String(typedE.message))
@@ -31,8 +41,13 @@ export class UserController {
         try {
             const { email } = req.body;
             const response = await UserService.getUserByEmail(email);
-            logger.log('info', 'api-UserController-getUserByEmail() | SUCCESS')
-            res.status(200).send(response);
+            if (response) {
+                logger.log('info', 'api-UserController-getUserByEmail() | SUCCESS')
+                res.status(200).send(response);
+            } else {
+                logger.log('info', 'api-UserController-getUserByEmail() | Error | User not found')
+                res.status(404).send({error: 'User not found'});
+            }
         } catch (e: unknown) {
             const typedE = e as Error
             logger.log('error', 'api-UserController-getUserByEmail() | Error | ' + String(typedE.message))
@@ -46,8 +61,12 @@ export class UserController {
             const user = await UserService.getUserById(id);
             if (user) {
                 const response = await UserService.updateUser(id, username, first_name, last_name);
-                logger.log('info', 'api-UserController-updateUser() | SUCCESS')
-                res.status(200).send(response);
+                if (response) {
+                    logger.log('info', 'api-UserController-updateUser() | SUCCESS')
+                    res.status(200).send(response);
+                } else {
+                    throw new Error('Error updating user');
+                }
             } else {
                 res.status(404).send({error: 'User not found'});
             }
@@ -60,19 +79,22 @@ export class UserController {
     public static async deleteUser(req: Request, res: Response) {
         try {
             const { id } = req.params;
-            const response = await UserService.deleteUser(id)
-            logger.log('info', 'api-UserController-deleteUser() | SUCCESS')
-            res.status(200).send(response)
-        } catch (e: unknown) {
-            if (e instanceof Error) {
-                const errorMsg = String(e.message)
-                logger.log('error', 'api-UserController-deleteUser() | ERROR | ' + errorMsg)
-                res.status(500).send({error: errorMsg})
-            } else {   
-                const errorMsg = String(e)
-                logger.log('error', 'api-UserController-deleteUser() | ERROR | ' + errorMsg)
-                res.status(500).send({error: errorMsg})
+            const user = await UserService.getUserById(id);
+            if (user) {
+                const response = await UserService.deleteUser(id);
+                if (response) {
+                    logger.log('info', 'api-UserController-deleteUser() | SUCCESS')
+                    res.status(200).send(response);
+                } else {
+                    throw new Error('Error deleting user');
+                }
+            } else {
+                res.status(404).send({error: 'User not found'});
             }
+        } catch (e: unknown) {
+            const typedE = e as Error
+            logger.log('error', 'api-UserController-deleteUser() | Error | ' + String(typedE.message))
+            res.status(500).send({error: typedE.message});
         }
     }
     public static async updateLastLogin(req: Request, res: Response) {
@@ -81,6 +103,12 @@ export class UserController {
             const user = await UserService.getUserById(id);
             if (user) {
                 const response = await UserService.updateLastLogin(id);
+                if (response) {
+                    logger.log('info', 'api-UserController-updateLastLogin() | SUCCESS')
+                    res.status(200).send(response);
+                } else {
+                    throw new Error('Error updating last login');
+                }
                 logger.log('info', 'api-UserController-updateLastLogin() | SUCCESS')
                 res.status(200).send(response);
             } else {
@@ -99,8 +127,12 @@ export class UserController {
             const user = await UserService.getUserById(id);
             if (user) {
                 const response = await UserService.setRole(id, is_admin);
-                logger.log('info', 'api-UserController-setRole() | SUCCESS')
-                res.status(200).send(response);
+                if (response) {
+                    logger.log('info', 'api-UserController-setRole() | SUCCESS')
+                    res.status(200).send(response);
+                } else {
+                    throw new Error('Error setting role');
+                }
             } else {
                 res.status(404).send({error: 'User not found'});
             }
